cambios en Student, details, edit, delete

+++ Controllers/StudentsController.cs
@@ -104,21 +104,23 @@ namespace ContosoUniversity.Controllers
                 return NotFound();
             }
 
-            var studentToUpdate = await _context.Students.FirstOrDefaultAsync(s => s.ID == id);
+            var studentToUpdate = await _context.Students.FirstOrDefaultAsync(student => student.ID == id);
 
             if (await TryUpdateModelAsync<Student> (
                 studentToUpdate,
                 "",
-                s => s.FirstName, s => s.LastName, s => s.EnrollmentDate))
+                s => s.FirstName, 
+                s => s.LastName, 
+                s => s.EnrollmentDate))
             {
                 try
                 {
                     await _context.SaveChangesAsync();
                     return RedirectToAction(nameof(Index));
-                } catch (DbUpdateException)
+                } catch (DbUpdateException e)
                 {
                     ModelState.AddModelError("", "Unable to save changes." +
-                        "Try again");
+                        "Try again. Error: " + e);
                 }
             }
 
@@ -126,7 +128,7 @@ namespace ContosoUniversity.Controllers
         }
 
         // GET: Students/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
         {
             if (id == null)
             {
@@ -134,10 +136,15 @@ namespace ContosoUniversity.Controllers
             }
 
             var student = await _context.Students
+                .AsNoTracking()
                 .FirstOrDefaultAsync(m => m.ID == id);
             if (student == null)
             {
                 return NotFound();
+            }
+            if (saveChangesError.GetValueOrDefault())
+            {
+                ViewData["ErrorMessage"] = "Delete failed";
             }
 
             return View(student);
@@ -149,9 +156,20 @@ namespace ContosoUniversity.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var student = await _context.Students.FindAsync(id);
-            _context.Students.Remove(student);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            if (student == null)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+
+            try           
+            {
+                _context.Students.Remove(student);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            } catch (DbUpdateException /* e */)
+            {
+                return RedirectToAction(nameof(Delete), new { id = id, saveChangesError = true });
+            }            
         }
 
         private bool StudentExists(int id)
